feat: bias_trasnforms.get_spatial_bc_quantiles()
Just trying to mimic the linear calibration. It's not clear the indices
used to select and slice the quantiles.

There is a weakness here since the quantiles are estiamted in a previous
step, it lacks some lock to guarantee that the choosed coefficients file
is the correct pair with the data to be corrected.