Time dependent QDM (#226)
* Initializing WindowedQuantileDeltaMappingCorrection

* feat: window_mask() method

A masking for measurements outside the intended time window of interest.

* feat: window_mask() method directly on QDM

A masking for measurements outside the intended time window of interest.
Building it directly in the standard QDM method.

* refact: run_single including bias time index

A requirements to estimate parameters by time window.

* feat: Quantiles estimated within moving time windows

This allows monthly quantiles or any other scales.

* refactor: local_qdm running on time windows

Expects the CDFs were estimated within time windows and apply the QDM
correction respecting those.

* feat: Method to guide periods along a year

This is used to support applying filters and masks.

* doc: Info on assuming nearest available CDF

* clean: Removing WindowedQuantileDeltaMappingCorrection

A different direction. Instead of that, I generatlized the standard
QuantileDeltaMappingCorrection to deal with N periods, which N=1 is the
same behavior of the original function.

* refactor: Using named arguments

Using named arguments allow more freedom on the functions signatures
thus more freedom for refactoring. Here we initiated the requirement on
time to be able to apply QDM on different time scales, such as
seasonally.

* Adding ipython to dev environment

* test: Adjusting tests to run with new temporal QDM

* cfg: Updating pixi.lock

* doc: window_mask()

* doc, style: _window_center()

* style:

* style:

* fix: Use nearest window center

* Replacing variable t by idt

As requested by @grantbuster

* refact: Renaming variables

As requested by @grantbuster

* refact, fix: window_mask doesn't need rouding

In case of a resolution higher than daily, it would make sense using the
window extent without rounding it. Also fix a sign when after 365.

* test: Testing window_mask()

* doc: Improving window_mask() documentation

* doc: Updating documentation to reflect new dimension on time window

Issue noticed by @grantbuster

* refact: Renaming time to time_index

As requested by @bnb32.

* doc: Information on the reference attributes

As requested by @grantbuster.